feat: ajustes functionalities

+++ src/modules/functionalities/functionalities.controller.ts
@@ -1,7 +1,6 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { FunctionalitiesService } from './functionalities.service';
 import { CreateFunctionalityDto } from './dto/create-functionality.dto';
-import { UpdateFunctionalityDto } from './dto/update-functionality.dto';
 
 @Injectable()
 export class FunctionalitiesController {
@@ -9,19 +8,23 @@ export class FunctionalitiesController {
     private readonly functionalitiesService: FunctionalitiesService,
   ) {}
 
-  async create(body: CreateFunctionalityDto) {
-    return this.functionalitiesService.create(body);
+  async createOrUpdate(body: CreateFunctionalityDto[]) {
+    return this.functionalitiesService.createOrUpdate(body);
   }
 
   async createMany(body: CreateFunctionalityDto[]) {
     return this.functionalitiesService.createMany(body);
   }
 
-  async update(id: number, body: UpdateFunctionalityDto) {
-    return this.functionalitiesService.update(+id, body);
-  }
-
   async remove(id: number) {
+    if (!id) throw new BadRequestException('Id não enviado.');
+    const isSomeTypeUsing =
+      await this.functionalitiesService.findOneTypesUsing(id);
+
+    if (isSomeTypeUsing)
+      throw new BadRequestException(
+        'Já existe um Tipo de usuário usando essa funcionalidade.',
+      );
     return this.functionalitiesService.remove(+id);
   }
 }
